5.4.0 server refactor
