+++ package.json
@@ -1,6 +1,6 @@
 {
 	"name": "odi",
-	"version": "5.3.5",
+	"version": "5.4.0-WIP",
 	"private": true,
 	"description": "Odi's program to autorun and interact !",
 	"main": "odi.sh",
